+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.5.0]
+
+ - comprehensive FDR implementation for differential expression analysis
+ - FDR-based visualization in `volcano_de` plots: support for local/tail FDR y-axes and coloring
+ - posterior tail probability for differential expression
+ - itroduction of "is\_de" boolean column in `adata.var` to indicate differential expression based on significance threshold
+ - more flexible `volcano_de` plot with FDR/PTP-based thresholding and y-axis options
+ - "signal" and "strength" columns in stringDB gene-set enrichment analysis
+ - expand testing
+
 ## [0.4.3]
 
  - make sure all components are packaged
@@ -18,9 +28,6 @@ All notable changes to this project will be documented in this file.
 
  - StringDBReport class for gene set visualization and reporting
  - make sure da directions categories are always retained and ordered correctly
- - comprehensive FDR implementation for differential expression analysis
- - FDR-based visualization in `volcano_de` plots: support for local/tail FDR y-axes and coloring
- - more flexible `volcano_de` plot with FDR-based coloring and y-axis options
  - `fold_change_mode` parameter for heatmap to only show fold-change instead of split tiles
  - implement `RunInfo` utility to fetch information about previous runs
  - bugfix passing `ax` to `kompot.plot.embedding`